Use settings for separate location for installing Go tools
